% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSize_Proportions.R
\name{EffectSize_Proportions}
\alias{EffectSize_Proportions}
\title{Determining required effect size with two proportions of equal n}
\usage{
EffectSize_Proportions(n = NULL, baserate = NULL,
  alternative = c("two.sided", "less", "greater"), treatmentrate = 1,
  sig.level = 0.05, power = 0.8)
}
\arguments{
\item{n}{The per-group sample size. Can be a single integer \code{n} or a combination of integers \code{c(n1, n2, etc)}.}

\item{baserate}{The base rate for the population and probability in question. Required for all analyses.}

\item{alternative}{The type of hypothesis. Can be \code{"greater"} (p1 > p2), \code{"less"} (p1 < p2), or \code{"two.sided"} (p1 != p2).}

\item{treatmentrate}{The proportion of those in the treatment group who are expected to complete treatment; see details.}

\item{sig.level}{The significance level of the hypothetical test. Defaults to \code{0.05}.}

\item{power}{The power of the hypothetical test. Defaults to \code{0.8}.}
}
\description{
This function runs power calculations with two proportions of equal n, determining what effect size is necessary to see a statistically significant effect. Results are printed and can also be assigned.
}
\details{
The effect size input is a "practical" effect size in that it is the
difference in observed probabilities, unlike Cohen's h. The calculations
adjust for treatment rate, so consider the effect size the mean effect among
those who successfully receive treatment.

The treatment rate is used to (crudely) calculate the effect size. Assuming
that those who do not receive treatment will be identical to the control
group, the required effect size will need to be inversely proportional to the
treatment rate. That is, if an effect of .1 is needed to see a significant
effect, and only 90% of the treatment group receives treatment, the required
effect size among those 90% treated becomes .1/.9 = .11.

"h" refers to Cohen's h, the formal score for effect size in tests of two
means. In the output, Cohen's h is calculated after taking treatment rate
into effect - it is the h for the full treatment group, including those who
do not receive treatment. If the treatment rate is lower than 1, h will be
lower.

Printed value is rounded; for an unrounded value, assign output to an object.
}
\examples{
Please refer to vignette for detailed examples.

#Determining effect size given n, etc
EffectSize_Proportions(n = 100, baserate = .5, alternative = "greater", treatmentrate = .8)

#Determining required effect size across multiple n's
EffectSize_Proportions(n = c(100, 200, 300), baserate = .5, alternative = "greater", treatmentrate = .8)
 
#Assigning (unrounded) required effect size
e <- EffectSize_Proportions(n = 100, baserate = .5, alternative = "greater", treatmentrate = .8)
e
}

